+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ## Added
+- Added Client ID setting (final name TBD)
 - Added ability to preview account IDs in element tables
 
 ## Changed
@@ -15,10 +16,10 @@
 - `Marketplace::getInstance()->accounts->createAccountLink($accountId, $params)` has been added, which may be what you’d prefer if you were using `createLoginLink()`
 
 ## Removed
-- OAuth Client plugin in favour of Social Login
+- OAuth Client removed as a dependency
 - Removed unused provider for standards Stripe accounts (only Express accounts were and are supported)
 - Removed internal references to potentially supporting direct charges; Marketplace is only going to support destination changes for the forseeable future.
-- Removed now non-existent app handle from settings, and `getAppHandle()` from handles service. Unlike OAuth Client, Social Login doesn’t have the concept of “apps.”
+- Removed now non-existent app handle from settings, and `getAppHandle()` from handles service. Unlike OAuth Client, the underlying auth plugin doesn’t have a concept of “apps.”
 - Removed display of orders by a payee on a payee’s user page
 - Removed support for configuring fees visually via the settings in the control panel